I have changed the name of AbstractClass and inserted the collide-method.
I have made changes so that the collide method from Ammunition and Bullet can be inserted.
