+++ WeaponPack/src/main/java/com/github/sef24sp4/weaponpack/machinegun/Bullet.java
@@ -5,10 +5,10 @@ import com.github.sef24sp4.common.entities.ICollidableEntity;
 import com.github.sef24sp4.common.entities.IEntity;
 import com.github.sef24sp4.common.interfaces.IEntityManager;
 import com.github.sef24sp4.common.metadata.GameElementType;
-import com.github.sef24sp4.weaponpack.WeaponDamageLevel;
+import com.github.sef24sp4.weaponpack.AbstractProjectile;
 
-public class Bullet extends WeaponDamageLevel implements ICollidableEntity, IAttackingEntity {
-	private final int damage = getDamageMachineGun();
+public class Bullet extends AbstractProjectile implements ICollidableEntity, IAttackingEntity {
+	private final int damage = 2;
 
 	/**
 	 * Constructs the BaseProjectile with the shooter.
@@ -17,16 +17,7 @@ public class Bullet extends WeaponDamageLevel implements ICollidableEntity, IAtt
 	protected Bullet(IEntity shooter) {
 		super(shooter);
 	}
-
-	//This method ensures, that when the projectile is removed, when a collision happens with otherEntities (except projectile and player).
-	@Override
-	public void collide(IEntityManager entityManager, ICollidableEntity otherEntity) {
-		if (this.getShooter() == otherEntity) return;
-		if (otherEntity.getType() == GameElementType.PROJECTILE) return;
-
-		entityManager.removeEntity(this);
-	}
-
+	
 	/**
 	 * Gets the amount of damage.
 	 * @return The value of damage.
